This PR:
```
BenchmarkInsertChain_empty_memdb-16             	    4736	    289978 ns/op	  177784 B/op	     504 allocs/op
BenchmarkInsertChain_empty_diskdb-16            	    4783	    268855 ns/op	  180592 B/op	     522 allocs/op
BenchmarkInsertChain_valueTx_memdb-16           	    3628	    346591 ns/op	  205057 B/op	     699 allocs/op
BenchmarkInsertChain_valueTx_diskdb-16          	    3615	    358547 ns/op	  208373 B/op	     716 allocs/op
BenchmarkInsertChain_valueTx_100kB_memdb-16     	    1489	    824494 ns/op	  743036 B/op	     697 allocs/op
BenchmarkInsertChain_valueTx_100kB_diskdb-16    	    1137	    970762 ns/op	 1051620 B/op	     726 allocs/op
BenchmarkInsertChain_uncles_memdb-16            	    3208	    406440 ns/op	  194773 B/op	     681 allocs/op
BenchmarkInsertChain_uncles_diskdb-16           	    3012	    393308 ns/op	  200140 B/op	     699 allocs/op
BenchmarkInsertChain_ring200_memdb-16           	     108	  10834430 ns/op	 4582279 B/op	   32858 allocs/op
BenchmarkInsertChain_ring200_diskdb-16          	     108	  11643432 ns/op	 4624940 B/op	   32199 allocs/op
BenchmarkInsertChain_ring1000_memdb-16          	      92	  13196275 ns/op	 5167392 B/op	   36919 allocs/op
BenchmarkInsertChain_ring1000_diskdb-16         	      91	  13142748 ns/op	 5206166 B/op	   36255 allocs/op
```
Previous:
```
BenchmarkInsertChain_empty_memdb-16             	    5588	    259871 ns/op	  177261 B/op	     504 allocs/op
BenchmarkInsertChain_empty_diskdb-16            	    6880	    290131 ns/op	  180113 B/op	     523 allocs/op
BenchmarkInsertChain_valueTx_memdb-16           	    3577	    331286 ns/op	  204962 B/op	     698 allocs/op
BenchmarkInsertChain_valueTx_diskdb-16          	    3609	    358991 ns/op	  208196 B/op	     715 allocs/op
BenchmarkInsertChain_valueTx_100kB_memdb-16     	    1407	    823753 ns/op	  743930 B/op	     696 allocs/op
BenchmarkInsertChain_valueTx_100kB_diskdb-16    	    1582	    951776 ns/op	 1049537 B/op	     733 allocs/op
BenchmarkInsertChain_uncles_memdb-16            	    3433	    380894 ns/op	  195124 B/op	     681 allocs/op
BenchmarkInsertChain_uncles_diskdb-16           	    2985	    370448 ns/op	  200189 B/op	     699 allocs/op
BenchmarkInsertChain_ring200_memdb-16           	     104	  11057091 ns/op	 4572293 B/op	   34810 allocs/op
BenchmarkInsertChain_ring200_diskdb-16          	     106	  11268588 ns/op	 4647128 B/op	   34139 allocs/op
BenchmarkInsertChain_ring1000_memdb-16          	      92	  13143731 ns/op	 5226717 B/op	   39041 allocs/op
BenchmarkInsertChain_ring1000_diskdb-16         	      78	  13567900 ns/op	 5265106 B/op	   38405 allocs/op
```