@MariusVanDerWijden 2020-10-06
better PR here: https://github.com/ethereum/go-ethereum/pull/19291